requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
websockets>=12.0
python-dotenv>=1.0.0
jinja2>=3.1.0

# PDF Generation
fpdf2>=2.7.0

# AI Design Review (R3 only)
dedalus-labs>=0.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Type checking (optional)
mypy>=1.7.0
